pyproject.toml
[tool]
[tool.poetry]
name = "stardate-goddard"
version = "0.0.1"
description = "Converts date time strings into Star Trek TNG star dates."
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
homepage = "https://github.com/Goddard/stardate"
authors = ["Ryein Goddard <ryein@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.6.2"
python-dateutil = "^2.8.2"

[tool.poetry.scripts]
stardate = 'stardate_goddard:main'

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
black = "^21.7b0"

[tool.black]
line-length = 80
